third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-border-color-shorthand-expected.txt

Tests that the border-color shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('border-color') is 'rgb(255, 0, 0) rgb(0, 0, 255) rgb(0, 128, 0) rgb(255, 255, 255)'
PASS computedStyle.getPropertyValue('border-color') is 'rgb(0, 0, 0)'
PASS computedStyle.getPropertyValue('border-color') is 'rgb(102, 102, 102)'
PASS computedStyle.getPropertyValue('border-color') is 'rgb(0, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
